PPPoE: code relocation
