+++ src/pppoe/Makefile
@@ -20,10 +20,9 @@
 #
 
 CC	 ?= gcc
-INCLUDES = -Iinclude -Ipppoe/include -I../lib -I../libbpf/src -I../libbpf/src/uapi
+INCLUDES = -Iinclude -I../include -I../../lib -I../../libbpf/src -I../../libbpf/src/uapi
 CFLAGS	 = -g -O2 -ggdb $(INCLUDES) -Werror -Wall -Wunused -Wstrict-prototypes -D_GNU_SOURCE -MMD
 #DEFS	 = -D_DEBUG_
-SUBDIRS   = pppoe
 DEFS	 =
 COMPILE	 = $(CC) $(CFLAGS) $(DEFS)
 
@@ -34,14 +33,8 @@ COMPILE	 = $(CC) $(CFLAGS) $(DEFS)
 SRCS=$(wildcard *.c)
 OBJS=$(SRCS:.c=.o)
 all: $(OBJS)
-	@set -e; \
-	for i in $(SUBDIRS); do \
-	$(MAKE) -C $$i || exit 1; done
 
 clean:
-	@set -e; \
-	for i in $(SUBDIRS); do \
-	$(MAKE) -C $$i clean || exit 1; done
 	@rm -vf *.a *.o *.d *~
 
 -include *.d
